+++ lib/src/builders/select/select_contracts.dart
@@ -11,9 +11,8 @@ abstract class SelectBuilderSql {
   });
 }
 
-abstract class SelectBuilderFetch {
-  Future<List<Map<String, dynamic>>> fetch();
-  Future<List<T>> fetchAs<T>();
+abstract class SelectBuilderFetch<T> {
+  Future<List<T>> fetch();
 }
 
 abstract class SelectBuilderAgregation {
@@ -24,7 +23,7 @@ abstract class SelectBuilderAgregation {
   Future<num> avg();
 }
 
-abstract class SelectBuilderLimit {
+abstract class SelectBuilderLimit<T> {
   /// Adds a LIMIT clause to the SQL query.
   ///
   /// [limitValue] The maximum number of rows to return.
@@ -38,10 +37,10 @@ abstract class SelectBuilderLimit {
   ///   .limit(10, 20) // Limit to 10 records, starting from the 21st record
   ///   .fetch();
   /// ```
-  SelectBuilderFetch limit(int limitValue, [int? offset]);
+  SelectBuilderFetch<T> limit(int limitValue, [int? offset]);
 }
 
-abstract class SelectBuilderJoin {
+abstract class SelectBuilderJoin<T> {
   /// Adds a JOIN clause to the query using the specified table and join matcher.
   ///
   /// The [table] parameter specifies the table to join with.
@@ -71,7 +70,7 @@ abstract class SelectBuilderJoin {
   /// ```
   ///
   /// Returns the [SelectBuilder] instance for method chaining.
-  SelectBuilder join(String table, JoinMatcher matcher);
+  SelectBuilder<T> join(String table, JoinMatcher matcher);
 
   /// Adds a raw JOIN clause to the query using the provided SQL string.
   ///
@@ -82,7 +81,7 @@ abstract class SelectBuilderJoin {
   /// The [raw] parameter is the raw SQL JOIN expression to add to the query.
   ///
   /// Returns the [SelectBuilder] instance for method chaining.
-  SelectBuilder joinRaw(String raw);
+  SelectBuilder<T> joinRaw(String raw);
 }
 
 enum OrderByType {
@@ -94,7 +93,7 @@ enum OrderByType {
   const OrderByType(this.value);
 }
 
-abstract class SelectBuilderOrderBy {
+abstract class SelectBuilderOrderBy<T> {
   /// Orders the query results by the specified fields.
   ///
   /// [fields] A list of field names to order by.
@@ -116,14 +115,17 @@ abstract class SelectBuilderOrderBy {
   ///   .orderBy(['department', 'last_name'], OrderByType.asc)
   ///   .fetch();
   /// ```
-  SelectBuilderFetch orderBy(
+  SelectBuilderFetch<T> orderBy(
     List<String> fields, [
     OrderByType type = OrderByType.desc,
   ]);
 }
 
-abstract class SelectBuilderWhere
-    implements SelectBuilderFetch, SelectBuilderLimit {
+abstract class SelectBuilderWhere<T>
+    implements
+        SelectBuilderFetch<T>,
+        SelectBuilderLimit<T>,
+        SelectBuilderOrderBy<T> {
   /// Adds a WHERE condition to the query using the specified field and matcher.
   ///
   /// This method is the primary way to filter results in a SELECT query by adding
@@ -160,7 +162,7 @@ abstract class SelectBuilderWhere
   /// @param matcher The matcher that defines the condition (Equals, Like, Between, etc.)
   /// @param type The type of boolean operator to use (AND or OR)
   /// @return The same builder instance for method chaining
-  SelectBuilderWhere where(
+  SelectBuilderWhere<T> where(
     String field,
     WhereMatcher matcher, {
     WhereMatcherType type = WhereMatcherType.and,
@@ -179,7 +181,7 @@ abstract class SelectBuilderWhere
   /// @param matcher The matcher that defines the condition to negate
   /// @param type The type of boolean operator to use (AND or OR)
   /// @return The same builder instance for method chaining
-  SelectBuilderWhere notWhere(
+  SelectBuilderWhere<T> notWhere(
     String field,
     WhereMatcher matcher, {
     WhereMatcherType type = WhereMatcherType.and,
@@ -197,7 +199,7 @@ abstract class SelectBuilderWhere
   /// @param raw The raw SQL expression to use in the WHERE clause
   /// @param type The type of boolean operator to use (AND or OR)
   /// @return The same builder instance for method chaining
-  SelectBuilderWhere whereRaw(
+  SelectBuilderWhere<T> whereRaw(
     String raw, {
     WhereMatcherType type = WhereMatcherType.and,
   });